+++ README.md
@@ -3,5 +3,5 @@
 ## Setup
 
 1. Clone this repo `git clone https://github.com/clarkdowner/four-seam.git`
-2. Install dependencies `npm install`
-3. Run it `npm run start`
+2. Install dependencies `yarn install`
+3. Run it `yarn build`
