--- app.ts
@@ -1,12 +1,0 @@
-import express from "express";
-
-const app = express();
-const port = 3000;
-
-app.get("/", (req, res) => {
-  res.send("Hello World!");
-});
-
-app.listen(port, () => {
-  console.log(`Four seam listening on port ${port}`);
-});
