+++ package.json
@@ -1,10 +1,19 @@
 {
   "dependencies": {
-    "express": "^4.18.2",
-    "typescript": "^5.1.3"
+    "express": "^4.18.2"
   },
+  "main": "src/index.ts",
   "scripts": {
-    "start": "npx tsc && node built/app.js"
+    "build": "webpack --progress --stats-error-details && node ./dist/bundle.js"
   },
-  "type": "module"
+  "type": "commonjs",
+  "devDependencies": {
+    "@types/express": "^4.17.17",
+    "@types/node": "^20.2.5",
+    "ts-loader": "^9.4.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3",
+    "webpack": "^5.85.1",
+    "webpack-cli": "^5.1.3"
+  }
 }
