+++ src/index.ts
@@ -1,0 +1,12 @@
+import express, {Request, Response} from "express";
+
+const app = express();
+const port = 3000;
+
+app.get("/", (req: Request, res: Response) => {
+  res.send("Hello World!");
+});
+
+app.listen(port, () => {
+  console.log(`Four seam listening on port ${port}`);
+});
