+++ tsconfig.json
@@ -1,10 +1,20 @@
 {
   "compilerOptions": {
-    "allowJs": false,
-    "moduleResolution": "Node",
-    "outDir": "./built",
+    "allowJs": true,
+    "allowSyntheticDefaultImports": true,
+    "allowUnreachableCode": false,
+    "esModuleInterop" : true,
+    "module": "ES2022",
+    "moduleResolution": "node",
+    "noFallthroughCasesInSwitch": true,
+    "noImplicitAny": true,
+    "noUnusedLocals": true,
+    "outDir": "./dist/",
+    "resolveJsonModule": true,
+    "sourceMap": true,
+    "strictNullChecks": true,
     "target": "ES2022"
   },
-  "include": ["./src/**/*", "app.ts"],
+  "include": ["./src/**/*"],
   "exclude": ["node_modules"]
 }
